# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Nemo V2
  version: 1.0.0
...
paths:
  /api/v1/platform/modules:
    get:
      tags:
        - FMS JS > Platform
      summary: Get Console Modules
      description: Get all platform modules available for console access
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Console permissions fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userid:
                        type: string
                        description: User ID
                      modules:
                        type: array
                        description: Array of platform modules
                        items:
                          type: object
                          properties:
                            moduleid:
                              type: string
                              description: Module ID
                            moduleinfo:
                              type: object
                              description: Module information including name, type, and other details
                              properties:
                                moduleid:
                                  type: string
                                  description: Module ID
                                modulename:
                                  type: string
                                  description: Module name
                                moduletype:
                                  type: string
                                  description: Module type (e.g., platform)
                                modulecode:
                                  type: string
                                  description: Module code
                                moduleinfo:
                                  type: object
                                  description: Additional module information
                                creditspervehicleday:
                                  type: number
                                  description: Credits per vehicle per day
                                isenabled:
                                  type: boolean
                                  description: Whether module is enabled
                  msg:
                    type: string
                    example: "Console permissions fetched successfully"
        "400":
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: string
                        example: "Invalid User ID format"
                  msg:
                    type: string
                    example: "Invalid input data"
        "403":
          description: Forbidden - Insufficient permissions
        "500":
          description: Internal Server Error - Get console permissions failed
  /api/v1/fms/account/modules:
    get:
      tags:
        - FMS JS > FMS > AccountMangement
      summary: Get Account Modules
      description: Get all web modules available for the account with their information
      security:
        - cookieAuth: []
      responses:
        "200":
          description: FMS permissions fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userid:
                        type: string
                        description: User ID
                      accountid:
                        type: string
                        description: Account ID
                      modules:
                        type: array
                        description: Array of available modules
                        items:
                          type: object
                          properties:
                            moduleid:
                              type: string
                              description: Module ID
                            moduleinfo:
                              type: object
                              description: Module information including name, type, and other details
                              properties:
                                moduleid:
                                  type: string
                                  description: Module ID
                                modulename:
                                  type: string
                                  description: Module name
                                moduletype:
                                  type: string
                                  description: Module type (e.g., web)
                                modulecode:
                                  type: string
                                  description: Module code
                                moduleinfo:
                                  type: object
                                  description: Additional module information
                                creditspervehicleday:
                                  type: number
                                  description: Credits per vehicle per day
                                isenabled:
                                  type: boolean
                                  description: Whether module is enabled
                  msg:
                    type: string
                    example: "FMS permissions fetched successfully"
        "400":
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: string
                        example: "Invalid User ID format"
                  msg:
                    type: string
                    example: "Invalid input data"
        "403":
          description: Forbidden - Insufficient permissions
        "500":
          description: Internal Server Error - Get FMS permissions failed
...
  /api/v1/platform/module/types:
    get:
      tags:
        - FMS JS > Platform > Module
      summary: Get Module Types
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/platform/module:
    post:
      tags:
        - FMS JS > Platform > Module
      summary: Create Module
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                modulename: MyTripInsights
                moduletype: web
                modulecode: mytripinsights
                creditspervehicleday: 5
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/platform/module/list:
    get:
      tags:
        - FMS JS > Platform > Module
      summary: List Modules
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/platform/module/{moduleid}:
    get:
      tags:
        - FMS JS > Platform > Module
      summary: Get Module
      parameters:
        - in: path
          name: moduleid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - FMS JS > Platform > Module
      summary: Update Module
      parameters:
        - in: path
          name: moduleid
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                modulename: Sample Module 2
                moduletype: api
                moduleinfo:
                  moduleurl: >-
                    https://dev-nemo.intellicar.io/uimodules/console/remoteEntry.js
                isenabled: false
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - FMS JS > Platform > Module
      summary: Delete Module
      parameters:
        - in: path
          name: moduleid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/platform/pkg/{pkgid}/modules:
    put:
      tags:
        - FMS JS > Platform > Packages
      summary: Update Package Modules
      parameters:
        - in: path
          name: pkgid
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                selectedmodules: []
                deselectedmodules:
                  - d3af9e7f-5d50-4bbb-9a38-f744787ff471
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...